update tugas validasi

+++ back-apps/app/Http/Controllers/StudentController.php
@@ -7,19 +7,36 @@ use Illuminate\Http\Request;
 
 class StudentController extends Controller
 {
-    //
-    public function index(){
+    public function index()
+    {
         $students = Student::all();
 
-        $data = [
-            'message' => 'Get all student',
-            'data' => $students
-        ];
+        if ($students->isNotEmpty()) {
+            $response = [
+                'message' => 'Successfully viewed all students',
+                'data' => $students,
+            ];
 
-        return response()->json($data, 200);
+            return response()->json($response, 200);
+        } else {
+            $response = [
+                'message' => 'No students found'
+            ];
+
+            return response()->json($response, 200);
+        }
     }
 
-    public function store(Request $request){
+    public function store(Request $request)
+    {
+        $request->validate([
+            'nama' =>'required|string|max:255',
+            'nim' =>'required|string|max:255|unique:student',
+            'email' =>'required|email|unique:student',
+            'jurusan' =>'required|string|max:255',
+
+        ]);
+
         $input = [
             'nama' => $request->nama,
             'nim' => $request->nim,
@@ -30,7 +47,7 @@ class StudentController extends Controller
         $student = Student::create($input);
 
         $data = [
-            'message' => 'Student is created succesfully',
+            'message' => 'Student created successfully',
             'data' => $student
         ];
 
@@ -39,32 +56,71 @@ class StudentController extends Controller
 
     public function update(Request $request, $id)
     {
-        $student = Student::findOrFail($id);
+        $student = Student::find($id);
 
-        $student->update([
-            'nama' => $request->nama,
-            'nim' => $request->nim,
-            'email' => $request->email,
-            'jurusan' => $request->jurusan
-        ]);
+        if ($student) {
+            $input = [
+                'nama' => $request->nama,
+                'nim' => $request->nim,
+                'email' => $request->email,
+                'jurusan' => $request->jurusan
+            ];
 
-        $data = [
-            'message' => 'Student is updated successfully',
-            'data' => $student
-        ];
+            $student->update($input);
 
-        return response()->json($data, 200);
+            $data = [
+                'message' => 'Student updated successfully',
+                'data' => $student
+            ];
+
+            return response()->json($data, 200);
+        } else {
+            $data = [
+                'message' => 'Student not found'
+            ];
+
+            return response()->json($data, 404);
+        }
     }
 
     public function destroy($id)
     {
-        $student = Student::findOrFail($id);
-        $student->delete();
+        $student = Student::find($id);
 
-        $data = [
-            'message' => 'Student is deleted successfully'
-        ];
+        if ($student) {
+            $student->delete();
 
-        return response()->json($data, 200);
+            $data = [
+                'message' => 'Student deleted successfully'
+            ];
+
+            return response()->json($data, 200);
+        } else {
+            $data = [
+                'message' => 'Student not found'
+            ];
+
+            return response()->json($data, 404);
+        }
+    }
+
+    public function show($id)
+    {
+        $student = Student::find($id);
+
+        if ($student) {
+            $data = [
+                'message' => 'Student details retrieved successfully',
+                'data' => $student
+            ];
+
+            return response()->json($data, 200);
+        } else {
+            $data = [
+                'message' => "Student with ID $id not found"
+            ];
+
+            return response()->json($data, 404);
+        }
     }
 }
